rename waitForStorage test
